+++ pkgs/shelf/lib/src/headers.dart
@@ -24,12 +24,24 @@ class Headers extends UnmodifiableMapView<String, List<String>> {
     } else if (values is Headers) {
       return values;
     } else {
-      return Headers._(values);
+      return Headers._from(values);
     }
   }
 
-  Headers._(Map<String, List<String>> values)
-      : super(CaseInsensitiveMap.fromEntries(values.entries
+  factory Headers.fromEntries(
+      Iterable<MapEntry<String, List<String>>>? values) {
+    if (values == null || (values is List && values.isEmpty)) {
+      return _emptyHeaders;
+    } else {
+      return Headers._fromEntries(values);
+    }
+  }
+
+  Headers._from(Map<String, List<String>> values)
+      : this._fromEntries(values.entries);
+
+  Headers._fromEntries(Iterable<MapEntry<String, List<String>>> entries)
+      : super(CaseInsensitiveMap.fromEntries(entries
             .where((e) => e.value.isNotEmpty)
             .map((e) => MapEntry(e.key, List.unmodifiable(e.value)))));